category_service/requirements.txt
azure-functions
requests
azure-cosmos